+++ NSTypes/Bit/src/DImage_Bit.cpp
@@ -141,7 +141,7 @@ void Image<Bit>::clone(const Image<Bit> &rhs)
 }
 
 template <>
-RES_T Image<Bit>::setPixel(UINT offset, const Bit &value)
+RES_T Image<Bit>::setPixel(size_t offset, const Bit &value)
 {
     this->lines[offset/width][offset%width] = value;
 }
